Fix commons errs
Prevent uninstall/--reset error

+++ lamw_manager/build-win-installer.sh
@@ -56,7 +56,10 @@ fi
 	printf "%b" "${file_str[*]}"
 
 	if [ -e "C:\Program Files (x86)\Inno Setup 5" ]; then
+		printf "\n\nBuild lamw_manager_setup.exe ...\n\n"
 		"C:\Program Files (x86)\Inno Setup 5\ISCC.exe" "$LAMW_MANAGER_PATH\lamw_manager.iss"
+		if [ -e "$LAMW_MANAGER_PATH\Output\lamw_manager_setup.exe" ]; then 
+			cp "$LAMW_MANAGER_PATH\Output\lamw_manager_setup.exe" ./
+		fi
 	fi
 
-	cp "$LAMW_MANAGER_PATH\Output\lamw_manager_setup.exe" ./
